requirements.txt
django
djangorestframework
environs
whitenoise
